# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: number;
  business_name: string;
  /** @nullable */
  business_website?: string | null;
  contact_email: string;
  /** @nullable */
  contact_phone?: string | null;
  /** @nullable */
  trade_type?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  google_rating?: string | null;
  /** @nullable */
  google_review_count?: string | null;
  /** @nullable */
  pagespeed_score?: string | null;
  status: string;
  created_at: string;
}

export interface LeadInput {
  /** @minLength 1 */
  business_name: string;
  /** @nullable */
  business_website?: string | null;
  contact_email: string;
  /** @nullable */
  contact_phone?: string | null;
  /** @nullable */
  trade_type?: string | null;
  /** @nullable */
  city?: string | null;
}

export interface LeadUpdate {
  status?: string;
  /** @nullable */
  google_rating?: string | null;
  /** @nullable */
  google_review_count?: string | null;
  /** @nullable */
  pagespeed_score?: string | null;
}

export interface StatGroup {
  label: string;
  count: number;
}

export interface LeadStats {
  total: number;
  by_status: StatGroup[];
  by_trade: StatGroup[];
  recent_count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListLeadsParams = {
status?: string;
trade_type?: string;
};