# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  /** @minLength 1 */
  business_name: string;
  /** @nullable */
  business_website?: string | null;
  contact_email: string;
  /** @nullable */
  contact_phone?: string | null;
  /** @nullable */
  trade_type?: string | null;
  /** @nullable */
  city?: string | null;
}