# lib/api-zod/src/generated/types/leadStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatGroup } from './statGroup';

export interface LeadStats {
  total: number;
  by_status: StatGroup[];
  by_trade: StatGroup[];
  recent_count: number;
}