# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadUpdate {
  status?: string;
  /** @nullable */
  google_rating?: string | null;
  /** @nullable */
  google_review_count?: string | null;
  /** @nullable */
  pagespeed_score?: string | null;
}