# lib/api-zod/src/generated/types/statGroup.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LeadForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatGroup {
  label: string;
  count: number;
}